DB query changes are done

+++ TCB.G1Group/DBQuerys/DBQuerys.sql
@@ -3,7 +3,7 @@ create schema if not exists public;
 create table users
 (
     id serial PRIMARY KEY,
-    telegram_client_id int,
+    telegram_client_id bigint,
     password           varchar(24),
     phone_number       varchar(13)
 );
@@ -12,11 +12,11 @@ create table users
 create table clients
 (
     id serial PRIMARY KEY,
-    user_id          int,
-    telegram_chat_id int,
+    user_id          bigint,
+    telegram_chat_id bigint,
     nickname         varchar(30),
     is_premium       bool,
-    status           int,
+    status           bigint,
     FOREIGN KEY (user_id)
         references users (id)
 );
@@ -25,9 +25,9 @@ create table anonym_chats
 (
     id          serial PRIMARY key,
     create_date date,
-    from_id     int,
-    to_id       int,
-    state       int,
+    from_id     bigint,
+    to_id       bigint,
+    state       bigint,
     FOREIGN KEY (from_id)
         REFERENCES clients (id),
     FOREIGN KEY (to_id)
@@ -39,8 +39,8 @@ create table boards
 (
     id       serial PRIMARY KEY,
     nickname varchar(30),
-    owner_id int,
-    board_status int,
+    owner_id bigint,
+    board_status bigint,
     FOREIGN KEY (owner_id)
         REFERENCES clients (id)
 );
@@ -49,11 +49,11 @@ create table boards
 create table messages
 (
     id       serial PRIMARY key,
-    from_id  int,
+    from_id  bigint,
     message  text,
-    chat_id  int,
-    type     int,
-    board_id int,
+    chat_id  bigint,
+    type     bigint,
+    board_id bigint,
     FOREIGN KEY (from_id)
         REFERENCES clients (id),
     FOREIGN KEY (board_id)
